// src/services/api/models/blockchain.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @enum {string}
 */
export enum Blockchain {
    Ethereum = 'Ethereum',
    Polygon = 'Polygon',
    Flow = 'Flow',
    Solana = 'Solana',
    Wax = 'Wax',
    BinanceSmartChain = 'BinanceSmartChain'
}