// src/services/api/models/blockchain-asset-metadata.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Blockchain } from './blockchain';
import { RedemptionStatus } from './redemption-status';
/**
 * 
 * @export
 * @interface BlockchainAssetMetadata
 */
export interface BlockchainAssetMetadata {
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BlockchainAssetMetadata
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof BlockchainAssetMetadata
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BlockchainAssetMetadata
     */
    updated?: Date;
    /**
     * 
     * @type {Blockchain}
     * @memberof BlockchainAssetMetadata
     */
    blockchain?: Blockchain;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    ownerAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    contractAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    tokenId?: string | null;
    /**
     * 
     * @type {RedemptionStatus}
     * @memberof BlockchainAssetMetadata
     */
    redemptionStatus?: RedemptionStatus;
    /**
     * 
     * @type {boolean}
     * @memberof BlockchainAssetMetadata
     */
    isTransferable?: boolean;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    videoUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    rarity?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    mintNumber?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BlockchainAssetMetadata
     */
    totalSupply?: string | null;
    /**
     * 
     * @type {any}
     * @memberof BlockchainAssetMetadata
     */
    data?: any | null;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof BlockchainAssetMetadata
     */
    attributes?: { [key: string]: string; } | null;
}